Attempt to extract class, refactor Right panel
