+++ dynamic-wallpaper/src/view/RightTextPanel.java
@@ -36,6 +36,7 @@ import javax.swing.JTextArea;
 import javax.swing.filechooser.FileNameExtensionFilter;
 
 import model.TimeList;
+import view.subpanel.RightBottomButton;
 
 /**
  * This class creates right/text panel, <br>
@@ -44,7 +45,7 @@ import model.TimeList;
  * @author Hung Huu Vu
  *
  */
-@SuppressWarnings({ "serial", "PMD.LawOfDemeter", "PMD.CommentSize"  })
+@SuppressWarnings({ "serial", "PMD.LawOfDemeter", "PMD.CommentSize" })
 //Ignore comment size (GPL copyright notice).
 public class RightTextPanel extends JPanel implements ActionListener {
 
@@ -84,43 +85,16 @@ public class RightTextPanel extends JPanel implements ActionListener {
   private static final JButton MY_X_LOG_BUTTON = new JButton("Clear LOG");
 
   /**
-   * Button to clear TIME LIST.
-   */
-  private static final JButton MY_X_TIME_BUTTON = new JButton("Clear TIME LIST");
-
-  /**
-   * Button to export TIME LIST.
-   */
-  private static final JButton MY_EXPORT_BUTTON = new JButton("Export TIME LIST"); // 09/14
-
-  /**
-   * Button to export TIME LIST.
-   */
-  private static final JButton MY_IMPORT_BUTTON = new JButton("Import TIME LIST"); // 09/14
-
-  /**
    * Sub panel for clear log button. Aesthetic purpose (avoiding layout component
    * resize).
    */
   private static final JPanel MY_X_LOG_PANEL = new JPanel();
-
+  
   /**
-   * Sub panel for clear time list button. Aesthetic purpose (avoiding layout
-   * component resize).
+   * Group of buttons at the bottom.
    */
-  private static final JPanel MY_X_TIME_PANEL = new JPanel();
-
-  /**
-   * Sub panel for export button. Aesthetic purpose (avoiding layout
-   * component resize).
-   */
-  private static final JPanel MY_EXPORT_PANEL = new JPanel(); // 09/14
-
-  /**
-   * Sub panel for import button. Aesthetic purpose (avoiding layout
-   * component resize).
-   */
-  private static final JPanel MY_IMPORT_PANEL = new JPanel(); // 09/14
+  // Extracting class to reduce code length and complexity.
+  private static final JPanel MY_BUTTON_SUBPANEL = new RightBottomButton();
 
   /**
    * LOG area label.
@@ -143,11 +117,6 @@ public class RightTextPanel extends JPanel implements ActionListener {
   private static final String MY_TIME_LIST_TEXT = "TIME_LIST";
 
   /**
-   * Cache the previous path for file chooser, both export and import buttons.
-   */
-  private static File myCachedDir = null;
-
-  /**
    * Constructor.
    */
   public RightTextPanel() {
@@ -162,13 +131,8 @@ public class RightTextPanel extends JPanel implements ActionListener {
     MY_TEXT_LOG.setEditable(false);
     MY_TEXT_TIME_LIST.setEditable(false);
     MY_X_LOG_BUTTON.addActionListener(this);
-    MY_X_TIME_BUTTON.addActionListener(this);
-    MY_EXPORT_BUTTON.addActionListener(this); // 09/14
-    MY_IMPORT_BUTTON.addActionListener(this); // 09/14
     MY_X_LOG_PANEL.add(MY_X_LOG_BUTTON);
-    MY_X_TIME_PANEL.add(MY_X_TIME_BUTTON);
-    MY_EXPORT_PANEL.add(MY_EXPORT_BUTTON); // 09/14
-    MY_IMPORT_PANEL.add(MY_IMPORT_BUTTON); // 09/14
+
     // Center aligned labels.
     MY_LOG_LABEL.setAlignmentX(CENTER_ALIGNMENT);
     MY_TIME_LABEL.setAlignmentX(CENTER_ALIGNMENT);
@@ -179,10 +143,12 @@ public class RightTextPanel extends JPanel implements ActionListener {
     add(MY_X_LOG_PANEL);
     add(MY_TIME_LABEL);
     add(SCROLL_TIME_LIST);
-    add(MY_EXPORT_PANEL);// 09/14
-    add(MY_IMPORT_PANEL);// 09/14
-    add(MY_X_TIME_PANEL);
+    add(MY_BUTTON_SUBPANEL);
 
+  }
+
+  public static String getTimeText() {
+    return MY_TEXT_TIME_LIST.getText();
   }
 
   /**
@@ -190,13 +156,18 @@ public class RightTextPanel extends JPanel implements ActionListener {
    *
    * @param theIndicator Indicate which text area is going to be changed: <br>
    *                     LOG for MY_TEXT_LOG (LOG message)<br>
-   *                     TIME_LIST for MY_TEXT_TIME_LIST (TIME LIST message)
+   *                     TIME_LIST for MY_TEXT_TIME_LIST (TIME LIST message) <br>
+   *                     RESET_TIME to empty MY_TEXT_TIME_LIST
    */
   public static void textSetter(final String theIndicator, final String theMsg) {
 
     if (theIndicator.equals(MY_LOG_TEXT)) {
 
       MY_TEXT_LOG.append(theMsg + "\n \n");
+
+    } else if (theIndicator.equals("RESET_TIME")) {
+
+      MY_TEXT_TIME_LIST.setText("");
 
     } else if (theIndicator.equals(MY_TIME_LIST_TEXT)) {
 
@@ -223,72 +194,6 @@ public class RightTextPanel extends JPanel implements ActionListener {
     if (theE.getSource() == MY_X_LOG_BUTTON) {
 
       MY_TEXT_LOG.setText("");
-
-    } else if (theE.getSource() == MY_X_TIME_BUTTON) {
-      // Clear time list (display message + data storage) when
-      // button is pressed.
-
-      // Reset time list data. (02/09)
-      controller.Controller.resetTimeList();
-
-      // Change requirement checklist.
-      NorthCheckListPanel.requirementSetter(20);
-
-      // Display log.
-      MY_TEXT_LOG.append("Clear time list successfully! \n" + "If the program is still running, \n"
-          + "please add new time to the list, \n" + "or stop the program. \n \n");
-
-      // Reset time list display/message.
-      MY_TEXT_TIME_LIST.setText("");
-
-    } else if (theE.getSource() == MY_EXPORT_BUTTON) {
-
-      exportButtonAction();
-
-    }
-
-  }
-
-  /**
-   * Action for export button. 09/14
-   */
-  private void exportButtonAction() {
-
-    final JFileChooser chooser = new JFileChooser();
-
-    // Only display txt files.
-    final FileNameExtensionFilter filter = new FileNameExtensionFilter("*.txt", "txt");
-    chooser.setFileFilter(filter);
-
-    if (myCachedDir != null) {
-      chooser.setCurrentDirectory(myCachedDir);
-    }
-
-    final int chosenOption = chooser.showOpenDialog(this);
-
-    // Holder of file path.
-    File savedFile = null;
-
-    if (chosenOption == JFileChooser.APPROVE_OPTION) {
-
-      // Open previous choosen path.
-      savedFile = chooser.getSelectedFile().getAbsoluteFile();
-
-      // Cache path.
-      myCachedDir = savedFile.getParentFile();
-
-      try {
-
-        // Save config to a file
-        PrintWriter writer = new PrintWriter(savedFile);
-        writer.println(MY_TEXT_TIME_LIST.getText());
-        writer.close();
-
-      } catch (FileNotFoundException e) {
-
-        System.out.println("Exception in RightTextPanel export writer.");
-
-      }
 
     }
